requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asteval==1.0.6
asttokens==3.0.0
attrs==25.3.0
autocommand==2.2.2
azure-ai-documentintelligence==1.0.1
azure-core==1.32.0
azure-identity==1.21.0
azure-storage-blob==12.25.0
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
cleo==2.1.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
courlan==1.3.2
crashtest==0.4.1
Crawl4AI==0.5.0.post4
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.5.0
dateparser==1.2.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docling==2.31.0
docling-core==2.27.0
docling-ibm-models==3.4.1
docling-parse==4.0.0
docopt==0.6.2
duckduckgo_search==8.0.0
dulwich==0.22.8
easyocr==1.7.2
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
firecrawl-py==1.14.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
html2text==2025.4.15
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
humanize==4.12.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
inflect==7.5.0
installer==0.7.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jusText==3.0.2
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.15
latex2mathml==3.77.0
lazy_loader==0.4
litellm==1.63.11
looseversion==1.3.0
lxml==5.3.1
lxml_html_clean==0.4.2
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
mistune==3.1.2
more-itertools==10.6.0
mpire==2.10.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.74.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pbs-installer==2025.4.9
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.22.0
pickleshare==0.7.5
pillow==10.4.0
pip-review==1.3.0
pipdeptree==2.26.0
pkginfo==1.12.1.2
platformdirs==4.3.6
playwright==1.50.0
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
portalocker==3.1.1
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pydot==3.0.4
pyee==12.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pymilvus==2.5.6
pymilvus.model==0.3.1
PyMuPDF==1.25.4
pyOpenSSL==25.0.0
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
pyperclip==1.9.0
pyproject_hooks==1.2.0
pystache==0.6.8
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytils==0.4.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.3.0
ragas==0.2.14
rank-bm25==0.2.2
RapidFuzz==3.12.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
semchunk==2.2.2
sentence-transformers==4.1.0
setuptools==76.0.0
shapely==2.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.3.3
stack-data==0.6.3
starlette==0.46.1
static==1.1.1
streamlit==1.44.1
streamlit-aggrid==1.1.4.post1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tld==0.13
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tools==0.1.9
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
traits==7.0.2
transformers==4.42.4
triton==3.2.0
trove-classifiers==2025.5.1.12
typeguard==4.4.2
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.33.1
unstructured_ingest==1.0.18
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarg==0.1.10
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
